Pull back ClauseDBIF to cdb/mod.rs
